fix(main): fixed incorrect max double which could be generated. On big numbers (> 20 Bilions) the number could not fit in the string and that may cause memory corruption. Added checks on json_buffer to prevent crashes and added constant MAX_JSON_DOUBLE which limits the max double which could be generated

+++ src/json_buffer/json_buffer.h
@@ -13,7 +13,7 @@ struct JsonBuffer{
   char* start_entries;
   char* start_suffix;
   formatter fmt_f;
-  int entry_len;
+  size_t entry_len;
   int num_entry;
   size_t cap;
 };
